# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: PaiHuo API
  version: 1.0.0
  description: PaiHuo control-plane contract. Browser sessions use an HttpOnly cookie obtained from POST /login.
servers:
  - url: /api/v1
...
paths:
...
  /artifacts:
    get:
      tags: [Artifacts]
      parameters:
        - {name: task_id, in: query, schema: {type: integer, format: int64, minimum: 1}}
        - {name: run_id, in: query, schema: {type: integer, format: int64, minimum: 1}}
      responses: {'200': {description: Artifact metadata list, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Artifact'}}}}}}
    post:
      tags: [Artifacts]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactInput'}}}}
      responses: {'201': {description: Artifact stored, content: {application/json: {schema: {$ref: '#/components/schemas/Artifact'}}}}, '400': {$ref: '#/components/responses/Error'}}
  /artifacts/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {tags: [Artifacts], responses: {'200': {description: Artifact metadata, content: {application/json: {schema: {$ref: '#/components/schemas/Artifact'}}}}, '404': {$ref: '#/components/responses/Error'}}}
    delete: {tags: [Artifacts], parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], responses: {'204': {description: Artifact metadata and unreferenced content deleted}}}
  /artifacts/{id}/content:
    get: {tags: [Artifacts], parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Artifact bytes}, '404': {$ref: '#/components/responses/Error'}}}
...
components:
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, description: Replays the first completed response for the same method and path, schema: {type: string, maxLength: 200}}
  responses:
    Error: {description: Error envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
...
    ArtifactInput:
      type: object
      required: [name, content]
      oneOf:
        - required: [task_id]
        - required: [run_id]
      properties:
        task_id: {type: integer, format: int64, minimum: 1}
        run_id: {type: integer, format: int64, minimum: 1}
        name: {type: string, minLength: 1}
        media_type: {type: string, default: application/octet-stream}
        content: {type: string, contentEncoding: base64}
        created_by: {type: string, default: operator}
        retention: {type: string}
    Artifact:
      type: object
      required: [id, name, media_type, content_hash, size, locator, created_by, created_at]
      properties:
        id: {type: integer, format: int64}
        task_id: {type: integer, format: int64}
        run_id: {type: integer, format: int64}
        name: {type: string}
        media_type: {type: string}
        content_hash: {type: string, pattern: '^[a-f0-9]{64}$'}
        size: {type: integer, format: int64, minimum: 0}
        locator: {type: string}
        created_by: {type: string}
        retention: {type: string}
        created_at: {type: string, format: date-time}